Move the config loader into `FlightConfig`
The `Loader` module uses `TTY::Config` to store configuration data. It
also controls all the file handling associated with reading and writing.

Instead it allows the object to be loaded and updated only. The update
method creates a new instance by design. This helps to ensure an up to
date version of the file is being used