Add an initializer method for __data__ core
There are some classes which need to set up the config core only if an
existing file does not exist. To facilitate this, the read method is
being moved into the Core itself.

The other half of this process is to run a dedicated initializer if it
is not a `read` action