Ensure configs can still read during a missing lock
Locking the file during `create` is a tricky task as there isn't a file
to be locked. Thus there is some underlining code that creates locks
(and deletes on failure) the placeholder file.

However during the lock, another process my try and read the placeholder
file. If this is the case, then it should ignore the files content