Start moving the path method onto the class
This is the first step at moving the path method. It will warn each
time a Config is used without moving the method.

The hack is required for compatibility with old Config classes which
will override the instance method. However the instance method must
work if the class method is defined